requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Cors==4.0.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# Redis and Background Tasks
redis==4.5.5
celery==5.3.6
flower==2.0.1

# HTTP and Requests
requests==2.31.0
urllib3==2.0.6

# Security and Authentication
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.4
PyJWT==2.8.0

# File Processing
python-magic==0.4.27
Pillow==10.0.1
python-dotenv==1.0.0

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Web Server
gunicorn==21.2.0
gevent==23.7.0

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
isort==5.12.0

# Monitoring and Logging
sentry-sdk[flask]==1.32.0

# API Documentation
flask-restx==1.2.0

# Form Handling
WTForms==3.0.1

# Date/Time
python-dateutil==2.8.2

# JSON Processing
jsonschema==4.17.3
# Environment Variables
python-decouple==3.8

# File Upload and Processing
Click==8.1.7

# Crypto and Hashing

# Graph and Network Analysis
networkx==3.3

# Character Encoding Detection
chardet==5.2.0

# System Monitoring
psutil==5.9.5

# Optional AI/ML cccc (commented out if not needed)
openai==0.28.1
anthropic==0.3.11
tiktoken==0.5.1

# Optional Blockchain Analysis (commented out if not needed)
# web3==6.10.0
# bitcoinlib==0.12.0

Flask-SocketIO~=5.5.1
binwalk~=2.3.3